No compiling intermediate commit of a ShapeList generative container.  Only the most LocalizableByDimCursor implemented but not yet tested.
ShapeList stores two synchronized arrays of Lists of Shapes and
Types.  That is, each Shape is associated with an actual Type
instance by index.  We are not using HashMaps because the order
matters for painting (which shape is top most), and search is
not required (yet).  The array index accumulates all dimensions
from 2, whereas x, y are expressed by the Shapes.