Fix ticket 800
When filling a region of interest with multiple threads, the filling was
incomplete.

This issue was due to relative vs absolute coordinates.

The startIndex was compared against the span[axis], but the span[axis] was
relative to origin[axis]. So let's rename startIndex to startOffset and
have it relative to the origin[axis], too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>