Update Bio-Formats version to the actual latest.
Using "5.0-SNAPSHOT" previously was just a placeholder, but now we are
tracking the correct latest version of Bio-Formats in the POMs, so
"4.3.3-SNAPSHOT" is better.

When we do a 2.0 beta release of ImgLib2, we will use a non-snapshot
version such as "4.3.3" to ensure stability in the entire stack of
dependencies.