Regional finder
This classes find the regional extrema of a 2 or 3 dimensional image. A
regional maxima is defined as follows:

"A regional maximum M of a grayscale image I is a connected components of
pixels such that every pixel in the neighborhood of M has a strictly lower
value."

The definition of a regional minimum is simply the opposite; the neighboring
pixels must be strictly brighter.

This class does not stipulate how much brighter a regional extreme must be from
the neighboring pixels, but only requires that it is brighter or dimmer. So a
connected component of pixels that is 1 higher in intensity from its neighbors
is treated the same as a connected component that is 100 higher in intensity
than it's neighbors.

The getRegionalExtrema() method returns an ArrayList of ArrayLists, where each
inner ArrayLists represents a regional extreme and contains the coordinates of
the pixels comprising that regional extreme.

Notably, this implementation does not allow for the identification of h-domes,
as mentioned by Luc Vincent in his paper, "Morphological Grayscale
Reconstruction in Image Analysis: Applications and Efficient Algorithms."

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>