Change all real operations to be based upon ComplexType
This facilitates reuse of code for Real and Complex image types.
Now an operation is not distinguished by the kinds of inputs it takes
but rather by whether it operates on the real or the complex fields
of those inputs. A real operation only operates on the real field of
a ComplexType.

Note - one constructor fell by the wayside here. RealInvert can no longer
take a no arg contructor. This may be an issue for Knip/Knime.