Move certain ImgLib1 tests into dedicated package
These problematic tests depend on ImageJ1, but the project to which they
belonged (imglib-algorithms) did not have such a test-scope dependency.

There were two options. Either: 1) add a test-scope dependency on
ImageJ1 to imglib-algorithms; or 2) create a new imglib-tests project
that has the necessary dependencies and houses these tests.

I chose solution #2, because it mirrors how we handled the same
situation in ImgLib2. In this way, we avoid all ImageJ dependencies
(even test-scope dependencies) except for in imglib-ij, imglib2-ij,
imglib-tests and imglib2-tests.