Use consistent version number for base POM
All of ImgLib2 is versioned as 2.0-SNAPSHOT, which is intuitive.

All of ImgLib1 is also versioned as 2.0-SNAPSHOT, which makes less
intuitive sense to me, but is fine.

However, the common base POM was versioned as 1.0-SNAPSHOT, which seems
rather silly. Since all subprojects are versioned at 2.0-SNAPSHOT, why
not have the base POM also versioned at 2.0-SNAPSHOT, and then have the
child POMs inherit the version number?