Change Jama dependency to compile scope
Jama is actually being used in the imglib2-realtransform project's main
code, not just in the test code. So the dependency cannot be test scope.
Unfortunately, Eclipse does not know the difference and just adds all
dependencies of both compile and test scope to its classpath. So the
only way to see this sort of issue is to build on the command line.