Add license-maven-plugin configuration
This makes the "mvn license:update-file-header" goal nearly usable for
stamping the copyright blurb in every source file header.

Still need to override the configuration for imglib2-algorithms-gpl, as
well as all the imglib1 projects.

This also changes the name of the license from "BSD" to "Simplified BSD
License" for clarity, and updates the pom-scijava version to 1.6, which
defines the version of license-maven-plugin.