Target Java 1.5 for imglib2 core
This creates a JAR with class version 49.0 instead of 50.0 (i.e.,
Java5-compatible instead of Java6-specific).

There will still be problems if Java6-specific API is used, but at
runtime, not compile time.

This change allows SCIFIO (which must continue to compile with Java5) to
depend on ImgLib2 at compile time without breakage. However, if you want
to actually *use* the ImgLib2-specific functionality, it may or may not
actually work, depending on whether the applicable portions of ImgLib2
use any Java6-specific features. But that is OK; anyone who wants to use
SCIFIO with ImgLib2 should be using Java6 anyway.

The only potential downside to this change is if the Java6 compiler
cannot make certain performance optimizations used at runtime by the
JIT. We would need to write some unit tests that measure performance, so
that we can compare class version 50.0 with 49.0.

For the time being, this compromise should allow SCIFIO to depend
cleanly on ImgLib2 core, without requiring a Java6 compiler in the OME
stack, or requiring Java5 compatibility in the ImgLib sources.