algorithm.pde: Gave up on optimizing 3D M.O.I.
I could not find a proper way to solve the fast diagonalization method
for 3x3 real symetric matrices. So I go back on using JTK for that.
Let it be known, that it was because of exhausted courage, not because
of impossiblity. A more courageous me will find in the future the
failing state in the previous commit.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>