added two static convenience methods to create Intervals.
FinalInterval.createMinSize( long ... ) takes the first half of its
parameters to specify the minimum, the second half to specify the
dimensions of the interval. For example, to create a 2D interval
from (10, 10) to (20, 40) use FinalInterval.createMinSize( 10, 10, 11, 31 ).

Similarly, FinalInterval.createMinMax( long ... ) takes the first
half of parameters as the minimum, the second half as the maximum.