Add new RegionsOfInterest: Angle and Line
These regions needed to create composite regions in IJ2 that correctly mirror
its overlay scheme.

In the process of adding these one version of move() was redefined
to be part of an abstract base class and removed from some implmentations.

Please note that LineRegionOfInteresrt has one TODO that needs to be fixed
before it is operational. The general line equation needs to be calculated
from the input points. This will follow shortly.