imglib2-core: Unsigned2BitType and Unsinged4BitType use a long index, and therefore can refer to up to Integer.MAX_SIZE * (64L / getBitsPerPixel()).
The question is what to do with methods int getIndex() and void updateIndex(int).
currently a cast to int is made.
These methods are never used for direct set/get of the index in the
underlying DataAccess in these 2 classes, and therefore could be using
long without issues, at least for ArrayImg derivatives.