Remove Bio-Formats dependency from imglib2-io
The imglib2-io project is licensed BSD-2, but Bio-Formats is GPL, so the
dependency is not allowed. The imglib2-io code will compile just fine
with only scifio.jar (the core of Bio-Formats), and does not need
bio-formats.jar as an explicit dependency. If a downstream project wants
to add support for PFFs via Bio-Formats, it is a simple matter of
including bio-formats.jar as a runtime dependency, rather than
inheriting it transitively from imglib2-io.