HyperSliceImgPlus: an hpuerslice view of an ImgPlus as an ImgPlus
This class offers access to a n-1-dimensional view of a source ImgPlus,
obtained by fixing a target dimension to a target position.
The source image is wrapped so there is data duplication.

Its result is exactly similar to the Views.hyperSlice(RandomAccessible, int, long)
static method, except that the returned class preserves the ImgPlus capabilities,
namely the metadata and iterability.

Internally, we strongly rely on the fantastic Views class
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>