Remove DiscNeighbohood and SphereNeighborhood
After sleeping over it, I decided that they are not good enough to be in
imglib2. They fill a very particular purpose (TrackMate calibrated units),
and they do not fit well with the neighborhoods design of Tobias.
Anote for modesty: I shall remember that my sh*t has a smell and that it
is rude to have others smell it.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>