OPS: big change in PointSet API (handling origins)
The existing use of setAnchor() and getAnchor() were inconsistent and
potentially buggy. They have been replaced with getOrigin() and
translate(byDeltas).

Old code copied array object refs to speed things up. But this was
not always the case and thus bugs could crop up. The new code eliminates
this inconsistency but at the cost of execution speed.