OPS: sandbox code for real coord sampling
This code allows one to define real coordinate samplings of regions of
space. It's an experiment that shows one can define them as grids, or
subsampling of a RegionOfInterest, or conditionally, etc.

There are numerous ways this could be extended (some documented in a
TODO in RegionSampling). I am committing this code so that it can be
referred to later and possibily used.