AbstractColorTable implements ArrayColorTable
AbstractColorTable already had a generic type indicating the type of
color table being returned. Now that there is an ArrayColorTable
interface, it makes more sense that AbstractColorTable implement the
most specific interface that it can, so that other consuming code can
take advantage of its generic parameter.

As part of this implementation a getValues() implementation was
provided.