Update copyrights and licenses for 2013
This was done using the commands:

    mvn license:update-project-license
    mvn license:update-file-header

This change also updates the authors list to include new authors, and
orders them according to total number of commits. (This is consistent
with before; it simply reorders everyone based on the new totals.)

Lastly, this change fixes the erroneous duplicate copyright blocks
previously present in many file headers.