Allow the ImgOpener to switch off file grouping
In this developers' performance tests, reading a single .tif from a
directory with 10,000 other files had a huge performance impact. It
seems that the TiffReader tries to read the whole directory every
time a slice is read, even if only one .tif file is read in the end.

This is a bit over the top, so let's add an option to switch that off,
keeping it on by default.

My gut feeling is that the default is wrong: most users will not care
about grouping files, and will not even know that ImgOpener tries to do
that! So let's discuss (between users, not between software architects
who do not use their own software on a daily basis) whether the default
should be changed, and if so, commit it as its own commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>