Let ImgOpener call SCIFIO's setAllowOpen(false)
SCIFIO has this option that it can prevent readers from opening a
stream just to determine whether they handle this type.

Of course, it would be better to just reuse the memory-mapped stream
because most readers need to inspect only the first couple of bytes
anyway.

This speeds up Tobias' benchmark from 317ms to 186ms.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>