Implement a numeric type interface hierarchy
This code is experimental and designed to investigate ways to extend the
numeric type system. In actuality it is a port of the Haskell type system as
it relates to numerics. Implemented in the OPS sandbox to avoid any clashes
elsewhere. Note that this code is for discussion use only. It can inform any
such discussions that may play out at the mini IJ2/Imglib hackathon in
Madison WI USA in Apr 2013.