Set the top-level's groupId/version correctly
There is a slight dichotomy between the way MiniMaven and Maven work:
MiniMaven does not re-set parent POMs because it wants a single tree
while Maven re-sets parent POMs according to what was parsed in the
pom.xml.

In this particular case, Fiji included the top-level ImgLib pom in its
pom-modules, but the top-level ImgLib pom.xml states that its parent is
pom-scijava, and core/pom.xml states that its parent is pom-imglib2. Due
to the way MiniMaven parsed things, pom-imglib2 thought that it is in
the sc.fiji group, though, and therefore imglib2-core did, too.

Fix this by setting the top-level groupId.

Unrelatedly, this is the correct thing to do anyway because otherwise,
Maven would not think that the top-level pom.xml defines the parent of
core/pom.xml (because the top-level project would now be in the group
org.scijava).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>