Remove generic parameter
Bind() is the only place where typing matters. Once again the set is
based on BitType. A Cursor<BitType> implementation is now in place.
In the short term there is no reason to use it. In the long run it
needs to be filled with probability values (not BitType) and perhaps
needs to be read only.