Add classes for unioning multiple typed spaces
The CombinedSpace provides a union of TypedSpaces.
The CombinedRealInterval provides a union of TypedRealIntervals.

The union is constructed by combining axes of the same AxisType.
In the case of the CombinedRealInterval, the realMin and realMax of each
dimension is computed by taking the lower and upper bounds,
respectively, of the constitent intervals' associated dimensions.