Add back deprecated delegators for the histogram classes/interfaces
ImgLib2 does not have the luxury of ImageJ2 of being in beta phase
where few people rely on the project, and those who do follow it
closely so they can easily fix backward-incompatible API changes.

Even worse: ImgLib2 is an important component in Fiji with quite a
few components depending on it.

As Fiji can only have *one* ImgLib2 version (if there were two on
the class path, one would invariably shadow the other), we *must*
retain backwards-compatibility at least for a few versions, even if
the feature in question was only meant to be in beta phase: if it
was available on 'master', we must consider it released and in use.

Therefore, this commit adds back (deprecated) delegators in the old
location of the histogram classes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>