Prevent the unit tests from opening a Java cup on MacOSX
On MacOSX, the unit tests which instantiate UI elements -- even if
they do not show them -- open a separate process with a separate dock
icon. This is pretty disruptive when developers run all the unit tests
in the background and try to do something else such as catching up
with their mail.

The apple.awt.UIElement property helps with that.

Suggested by David Philip Marshall in

	http://mid.gmane.org/1374227295612-5004030.post@n6.nabble.com

This commit was "borrowed" from https://github.com/imagej/imagej's
d347d36679d7aabb87c54321398189de2767e179, with an additional adjustment:
due to the (for a unit test) insane memory requirements, we need to adjust
the heap size in the argLine to 512m (128m is not enough).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>