ImagePlusAdapterTest: fix axis type comparison
This fixes the testDimensionality unit test to pass again.

The problem was that previously, querying the "axis" actually returned
just an AxisType. But after the merger of the 'img-metadata' branch, an
"axis" really is an Axis, which (if it's a TypedAxis) *has* an AxisType,
but *isn't* an AxisType. You have to ask for the type by calling type().

The bug was masked at compile time by the use of the equals method
(which accepts any Object) rather than the equality operator (==).