Provide support for nonlinear calibrations
The calibration code in the img-metadata package no longer defines
calibration(int d) and setCalibration(double v, int d). There is no
assumption in the axis definitions that calibration is linear.

DefaultCalibratedAxis has gone away. LinearAxis is now the usual default.

There are places in the code that make linear assumptions. They are marked
with TODOs.