Morphology: fix probable problems with missing iterated value
We need a starting value to compare to when computing the max over a
neighborhood.
For RealType and co, everything is fine: we can rely on
RealType.getMinValue(). For the more generic Type, hoping that we will
ALWAYS iterate over some value of the source image is bad faith. There
exists strels that do not iterate on any part of the source image at
some locations (typically borders). Therefore we must have a default
value to start with. To be ok with the definition of dilation, this
default value must be smaller than any of the pixel in the source, but
in the generic case, the user as to provide it.

A bit unsatisfactory.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>