LinearAxis: split interface and implementation
The reason is that we have a special LinearSpace, and in several places
we check whether an axis instanceof LinearAxis. So if LinearAxis is not
an interface we hamstring extensibility here. Perhaps ideally all the
various Axis implementations should be interfaces, but it is less
critical than for LinearAxis at this juncture.

There is also an issue with a circular package dependency between
net.imglib2.meta and net.imglib2.meta.axis:

- Both CalibratedSpace and CalibratedAxis (a class dependency of
  CalibratedSpace) are located in net.imglib2.meta.
- LinearSpace is also in net.imglib2.meta, but LinearAxis (a class
  dependency of LinearSpace) is located in net.imglib2.meta.axis.

So the design is clearly inconsistent, package-wise. However, the
consequences to consuming code are minimal from a practical perspective,
so rather than repackage a whole bunch of classes to avoid any hint of
circular package dependencies, we leave things as they are for now.