DefaultCalibratedRealInterval: fix bug during ctor
We avoid calling instance methods during construction, in case
subclasses override those methods in such a way that they do not work
yet when called from the parent class.

In this case, we have all the information we need (i.e., the number of
dimensions) and can pass that information to the assignDefaultAxes
method, such that calling numDimensions() becomes unnecessary.