DefaultUnitService: improve parsing error handling
It is bad to catch Exception because that will catch many other types of
exceptions including all RuntimeExceptions. It is also bad to completely
eat the exception with no way of knowing what it was.

We remedy both problems by:

A) Emitting the exception as an error using the LogService; and
B) Centrally catching each type of checked exception individually.