Added several new utility classes for Img/Labeling
Several new utility classes for Img/ImgPlus and Labeling were added.
Additionally a the new interface WrappedImg was added.

WrappedImg is an interface which can be used to retrieve the underlying
Img in e.g. a ImgPlus, as you need access to the Img to get detailed
information about it (e.g. is it an array img etc).

The helper classes ImgView, ImgPlusView and LabelingView allow one to wrap
an arbitrary RandomAccessibleInterval which than acts as an
Img/ImgPlus/Labeling.