CombinedSpace, Interval: improve thread safety
The update method of both CombinedSpace and CombinedRealInterval clears
the associated data structure of that class, but does not block access
to these data structures. This can lead to AIOOB and NPE exceptions
being thrown if any of the accessor methods are called in the middle of
an update.

To resolve this, both classes had their update methods placed in a
synchronized block, and all data structure accessors now go through a
synchronized block as well, to ensure these classes can not read and
write simultaneously.