SubpixelLocalization: do not update position if Matrix is singular.
Jama, on which we rely for the quadratic offseting, throws a
RuntImeException if the hessian Matrix is singular. This happens in some
cases.
Because the SubpixelLocalization generates threads to process its
tasks, these exception cannot be caught.
This commit adds a check to ensure we do the offset upate ONLY id the
matrix is not singular.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>