POM: update developers and contributors
According to the Maven docs, the "developers" section is supposed to be
for active developers. Rule of thumb: people you would contact with
questions about the project. The "contributors" section is for people
who contribute(d) to the project, but who are not officially responsible
for it. So a few of the people formerly tagged as "developers" are more
properly listed under "contributors" instead.