Remove imglib2-ops project
It migrated into the ImageJ OPS project as deprecated code:
https://github.com/imagej/imagej-ops/commit/a332184e6014567f47a14276c419f4ac3a0172f6

The goal is to fully replace the functionality of ImgLib2 OPS with that
of ImageJ OPS, according to its new and improved paradigm. And in the
meantime, those needing the old codebase can add a dependency on
net.imagej:imagej-ops to retain full backwards compatibility.