Remove imglib2-algorithms-legacy
In accordance with the discussion at the October 2014 ImgLib2 hackathon
in Madison, this project is being retired. It is not in active use as
far as we know. For anyone still using it: the
net.imglib2:imglib2-algorithms-legacy:2.0.0-beta-26 release will
continue to be available as a stable dependency. If you need to make
code changes to these classes for your project, the suggested approach
is to resurrect them from just prior to this commit, and fork them into
your project with a renamed package.