Synchronize bit type setter again
The previous version (in ByteArray, before ffeee80) ensured a synchronized
write access to the underlying array. Unsynchronizing unfortunately broke
ImageJ OPS, therefore we need to synchronize it again.

In a discussion about alternatives, Adrian Daerr suggested that
an alternative this developer suggested that used so-called Double-Checked
Locking does not work reliably. Example:

        newValue = array[i] | 1;
	array[i] = newValue;
        if (array[i] != newValue) {
                synchronized (array) {
			newValue = array[i] | 1;
			array[i] = newValue;
		}
        }

Let's assume that thread Albert is in the synchronized block, having just
read the very same array element as thread Bene, but Bene is just
modifying the value as per the second line. The Java Memory Model states
that Bene's modification is not necessarily seen, nor blocked by Albert's
synchronized write access because Bene is unsynchronized. And Albert's
modification will not necessarily be seen by Bene in the if() statement,
either – because Bene is unsynchronized.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>