Mark lightweight data objects with a common iface
The concept of a "view" is in several places across ImgLib2;
it is convenient to have a them all share a common interface.

Firstly, this allows for runtime introspection, so that programs
can decide whether it is warranted to make a copy of the data
(i.e., to "burn it in"), to avoid:

1. Costly recomputation when the lightweight data is accessed; and
2. Corrupting underlying data when the lightweight data mutates.

Secondly, this interface provides a common point for exploring the
ImgLib2 API, so that programmers can easily browse which
lightweight data structures exist, and how they are implemented.