ExtendedRandomAccessibleInterval: extend Interval
More specifically, an ExtendedRandomAccessibleInterval is now a
RandomAccessibleInterval, not just a RandomAccessible, as one probably
expects.

ExtendedRandomAccessibleInterval was not an Interval because its
OutOfBoundsStrategy defined its values outside the bounds of its source
Interval. I think the original idea was that this class lets an Interval
"transcend" its bounds and therefore no longer be limited by them.

But actually, there is no reason not to still implement the Interval
interface and continue to report those same bounds -- since it is not a
requirement of Intervals that they be undefined outside their bounds.

This is especially convenient to eliminate the common pattern of
extending an RAI using code like:

  extendedRAI = Views.interval(Views.extend(rai, oobs));

It is now sufficient to write simply:

  extendedRAI = Views.extend(rai, oobs);