Integer ShearTransforms with configurable shearing factor
Current ShearTransform only allows n-dimensional image to be sheared along the shear dimension, with respect to the reference dimension, by an integer factor of 1. This PR refactors ShearTransforms to allow transform with any integer shearFactor.

Views have also been updated to include static methods that allow a shearFactor to be specified.