Allow passing a type (or Supplier) to ImgFactory
Each ImgFactory instance is, according to the generics, bound to a
specific type. It is redundant to feed a type instance to the instance
methods in addition to that bound. Let's give the type instance (or
supplier of such) to the constructor up front, and introduce create
methods which do not require another explicit type instance.

All old constructors and methods are left in place but deprecated.
Old usages should still function. The only thing which will not work
is calling a deprecated typeless constructor, followed by one of the
new typeless create methods, since the type will be null in that case.

See #91.