Add constant OutOfBounds extension for unbound generics
`OutOfBoundsConstantValue` is bound to `T extends Type< T >` as the provided extension is copied. This does not allow to extend arbitrary generics with a constant value.
I assume that the rationale for that behavior is to avoid accidental modifications of the provided type, which could actually represent meaningful data.

To get around this restriction, I add a replication of `OutOfBoundsConstantValue` and `OutOfBoundsConstantValue` that uses a `Supplier< T >` instead of a `T` to extend the `RandomAccessibleInterval` with a constant value`.