WIP: Add method Casts.unchecked
Java has a very half backed implementation for generics. It is therefor
often necessary to perform unchecked casts.
The cumbersome syntax in Java forces the developer to either:
 * ignore the warnings, or
 * clutter the code with @SuppressWarnings annotations.

This commit suggest a third way, it introduce a nice and small method for
unchecked casts. And shows how it can be used to increase code readability.