Deprecate methods Views.offset
The methods Views.offset are supposed to be used to remove an offset from
an image. A good name for them would be Views.removeOffset.
The name View.offset is confusing, as it looks like an offset is added.

The methods, are deprecated because they should not be used, as their
naming confuses. Views.translate or Views.zeroMin can be used instead.