Add classes Parallelization and TaskExecutor
TaskExecutor should be used in image processing algorithms to
realize parallization instead of ExecutorService.
TaskExecutor has a single threaded implementation. That runs
with no parallelization overheaded, and it is easier to use
than ExecutorService.

The Parallelization class allows to set one TaskExecutor per thread.
This means that algorithms don't need to have a TaskExecutor argument.
Instead an algorithm can be called like this:

Parallelization.runSingleThreaded( () -> myAlgorithm() );

or

Parallelization.runMultiThreaded( () -> myAlgoritm() );