Fix Util.getArrayOrCellImgFactory to use Integer.MAX_VALUE - 8 as threshold
The method Utils.getArrayOrCellImgFactory failed to return a proper
ImgFactory for images of the size Integer.MAX_VALUE. This is because
the actual maximum possible java array size, depends on the JVM and
is slightly below Integer.MAX_VALUE.

We use MAX_ARRAY_SIZE = Integer.MAX_VALUE - 8.
This is similar to OpenJDK8's ArrayList.