Add IntegerTypes utility class
This commits adds a simple utility class to retrieve an IntegerType based on given min and max values, or on a max value and a flag for signed/unsigned.

It includes a test for all possible return types.

Also removed hard-coded version pinning in the test dependencies, and added myself as contributor.