BREAKING CHANGE: relax type constraint for converter targets
This requires that converting samplers and accessibles hold a Supplier
for conversion targets instead of a Type instance that could be used as
a Supplier.  The existing Converters convenience methods with Type
instances are retained such that old code using these methods compiles
with this version.  However, the implementation of samplers and
accessibles changed and may cause binary incompatibility