Default implement RealCursor.next()
The argument eagainst doing would potentially be performance: Every call to
the default next() goes	to polymorphic fwd() and get() calls.

I did a benchmark, and it doesn't matter. Probably because the default
implementation is so small it is inlined everywhere anyway. Additionally,
we had already implemented next() in	AbstractCursor, AbstractCursorInt,
etc, so we would have this problem already anyway.