WIP: fix compile errors with IterableRealInterval.getType()
IterableRealInterval and RandomAccessible cannot both have default
implementations, otherwise we run into a bug with generic bounds:
https://bugs.openjdk.org/browse/JDK-7120669

And for the same reason, the getType() method must be inherited from a
common super-type. Possibly it is possible to avoid this by not having
a default implementation at all. Something to try. On the other hand, a
Typed<T> interface is maybe a good idea?