Add recursive generic to RaView.
This is necessary to make apply(Function) work correctly for both, RaView
and RaiView. Effectively, the function given to RaView.apply() should
take RaView (or any parent, e.g., RandomAccessible). The function given to
RaiView.app;y() should take RaiView (or any parent, e.g., RandomAccessibleInterval).

I can find no other way to make this work.
Hopefully this is not too inconvenient because RaView is not supposed to be
used as type of variables, etc. It is just a construct to make the view-
chaining API possible
chains