Make TempArray work for volatile types
New method TempArray,forType(NativeType) creates a new instance
taking into account the underlying primitive type and whether the
NativeType is Volatile.

Vor volatile types, the T of TempArray<T> will *not* be a primitive
array type (as it is for non-volatile types). Instead, T is
VolatileArray<P> where P is the primitive array type.

VolatileArray<P> holds a P primitive array for data and a byte[]
primitive array for validity.