Add ScaledBrightnessARGBConverter
This is to enable in BDV a better display control for label images

This converter allows for better controling the visibility of ARGB
segmentation. The scale value is used to alter the brightness of
colors, which is better suited than altering R G and B in bulk.

This could be used in the BDV, to finely control the visibility of a
segmentation overlay in the shape of an ARGB image.